feat: 广告业务完成

+++ haoke-service/itcast-haoke-manage/itcast-haoke-manage-dubbo-server/itcast-haoke-manage-dubbo-server-house-resources/itcast-haoke-manage-dubbo-server-house-resources-dubbo-service/src/main/java/cn/itcast/haoke/dubbo/server/service/impl/HouseResourcesServiceImpl.java
@@ -61,7 +61,7 @@ public class HouseResourcesServiceImpl extends ServiceImpl<HouseResourcesMapper,
         IPage<HouseResources> page = this.page(new Page<HouseResources>(pageNum, pageSize), queryWrapper);
 
         // 返回分页数据
-        return new PageInfo<HouseResources>(Long.valueOf(page.getTotal()).intValue(), pageNum, pageSize, page.getRecords());
+        return new PageInfo<HouseResources>(page.getTotal(), pageNum, pageSize, page.getRecords());
     }
 
     /**
